[RCM] Replace the remote config agent client w/ an RFC compliant client
We had an existing remote config client for agent sub-processes, but in
the time since we first wrote that client, an RFC was published for
integrating with remote configuration in tracers and agent
subprocesses. That RFC introduced changes that made our existing client
non-compliant. This work transforms the client to match the RFC.

The client is split up across two packages. The "agent client" lives in
"pkg/config/remote" and consists of the networking layer that builds
requests and sends them over gRPC to the core agent. It leverages the
"state" package in "pkg/remoteconfig/stae" to process those updates
and store the state using the `Repository` type. This `Repository` type
can be utilized in the Go tracer client as well to prevent that work
from having to be implemented again.

The reason for the split is mostly that we're wanting to slowly migrate
remoteconfig code out of "pkg/config/remote" and into
"pkg/remoteconfig" for clarity. This lays the groundwork for doing that.

One important note is that the Go client technically goes further than
the RFC mandates with respect to TUF validation in order to make sure
that we can expand the RFC to support partial TUF validation down the
road.

Another important note is that the model for relaying updates to
interested product subsystems has changed. Previously we would create
channels to distribute updates, we now use a simple function callback
model. If an update is expensive to apply, the product subsystem is free
to setup a channel/goroutine pipeline and use their callback function to
schedule the work. This allows for more flexibility and keeps us from
forcing a product subsystem to use channels and goroutines.