Don't expire contexts for counters we may want to add later
Agent keeps reporting counters for a fixed period of time after the
client app has stopped sending them. Contexts are expired on a
different schedule. When context expiration time is low enough,
granularity of flushes becomes a hindrance, as counters sometimes are
sampled not frequently enough to keep their contexts alive.

This patch prevents active counter contexts from being expired,
regardless of their timestamp in the context resolver. Once the
reporting logic is done with the context, it can be expired using the
regular process though.